FireAlarm basics implemented

+++ src/fireAlarms/SensorInterface.java
@@ -3,8 +3,8 @@ package fireAlarms;
 //Sensor interface for every available reading sensors;
 interface  SensorInterface {
 
-	public double getTemperature();
-	public double getBatteryLevel();
+	public int getTemperature();
+	public int getBatteryLevel();
 	public int getSmokeLevel();
 	public int getCo2Level();
 }
